refactor(chat): extract Tab context UI and provider UI modules
Move external-context chips, context summary/permission labels, relevant
notes wiring, usage refresh, and provider UI gating out of Tab.ts into
tabContextUI and tabProviderUI. Keeps Tab lifecycle/init intact and
re-exports the public provider availability helpers for existing callers.
